refactor(crane-mcp): split skill-audit.ts — extract pure helpers to skill-audit-helpers.ts
skill-audit.ts was 502 LOC mixing pure helpers (frontmatter parsing, git
staleness, filesystem discovery) with executor/formatter logic. Split at
the stateless-vs-stateful seam:

- skill-audit-helpers.ts (new, 148 LOC): parseFrontmatter, gitLastTouched,
  daysSince, discoverSkills, findConsoleRoot, readSkillContent, Frontmatter,
  DiscoveredSkill — no API calls, no side effects
- skill-audit.ts (reduced, 355 LOC): schema, output types, executeSkillAudit,
  runSkillAuditAsync, runSkillAudit, formatReport; barrel re-exports helpers
  so all existing callers remain unchanged
